Removed News logic, update Races
