+++ ui-web-react/src/components/races/race-list.container.ts
@@ -1,5 +1,5 @@
 import { StoredState } from "../../model/types/datatypes";
-import { RaceListComponent } from "./race-list.component";
+import RaceListComponent from "./race-list.component";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
 import { racesRequested } from "../../model/actions/actions";
