+++ ui-web-react/src/model/actions/actions.ts
@@ -1,8 +1,6 @@
 import { AnyAction } from "redux";
 import {
-  NewsItem,
   RaceItem,
-  NewsItemExt,
   RaceItemExt,
   UserInfo,
   Alert,
@@ -13,14 +11,8 @@ import Optional from "optional-js";
 // ----------------------------------------------------------------------
 // Action lables
 // ----------------------------------------------------------------------
-export const NEWS_REQUESTED = "NEWS_REQUESTED";
-export const NEWS_LOADED = "NEWS_LOADED";
-
 export const RACES_REQUESTED = "RACES_REQUESTED";
 export const RACES_LOADED = "RACES_LOADED";
-
-export const SELECTED_NEWS_REQUESTED = "SELECTED_NEWS_REQUESTED";
-export const SELECTED_NEWS_LOADED = "SELECTED_NEWS_LOADED";
 
 export const SELECTED_RACE_REQUESTED = "SELECTED_RACE_REQUESTED";
 export const SELECTED_RACE_LOADED = "SELECTED_RACE_LOADED";
@@ -43,21 +35,9 @@ export const RACER_PROFILES_REQUEST_FAILED = "RACER_PROFILES_REQUEST_FAILED";
 // ----------------------------------------------------------------------
 // Actions classes
 // ----------------------------------------------------------------------
-export type NewsRequestedAction = AnyAction;
-export type NewsLoadedAction = AnyAction & {
-  items: Optional<NewsItem[]>;
-};
-
 export type RacesRequestedAction = AnyAction;
 export type RacesLoadedAction = AnyAction & {
   items: Optional<RaceItem[]>;
-};
-
-export type SelectedNewsRequestedAction = AnyAction & {
-  id: number;
-};
-export type SelectedNewsLoadedAction = AnyAction & {
-  newsItemExt: NewsItemExt;
 };
 
 export type SelectedRaceRequestedAction = AnyAction & {
@@ -91,29 +71,12 @@ export type RacerProfilesRequestFailedAction = AnyAction;
 // ----------------------------------------------------------------------
 // Actions
 // ----------------------------------------------------------------------
-export const newsRequested = (): NewsRequestedAction => ({
-  type: NEWS_REQUESTED
-});
-export const newsLoaded = (items: NewsItem[]): NewsLoadedAction => ({
-  type: NEWS_LOADED,
-  items: Optional.of(items)
-});
-
 export const racesRequested = (): RacesRequestedAction => ({
   type: RACES_REQUESTED
 });
 export const racesLoaded = (items: RaceItem[]): RacesLoadedAction => ({
   type: RACES_LOADED,
   items: Optional.of(items)
-});
-
-export const selectedNewsRequested = (id: number): SelectedNewsRequestedAction => ({
-  type: SELECTED_NEWS_REQUESTED,
-  id: id
-});
-export const selectedNewsLoaded = (newsItemExt: NewsItemExt): SelectedNewsLoadedAction => ({
-  type: SELECTED_NEWS_LOADED,
-  newsItemExt: newsItemExt
 });
 
 export const selectedRaceRequested = (id: number): SelectedRaceRequestedAction => ({
