+++ ui-web-react/src/model/reducers/reducers.ts
@@ -1,7 +1,6 @@
 import { combineReducers } from "redux";
 import { StoredState } from "../types/datatypes";
 import { userReducer } from "./user.reducer";
-import { newsReducer, selectedNewsReducer } from "./news.reducer";
 import { racesReducer, selectedRaceReducer } from "./race.reducer";
 import { alertsQueueReducer } from "./alertsQueue.reducer";
 import { racerProfilesReducer } from "./racerProfiles.reducer";
@@ -9,8 +8,6 @@ import { racerProfilesReducer } from "./racerProfiles.reducer";
 const raceLogAppState = combineReducers<StoredState>({
   user: userReducer,
   racerProfiles: racerProfilesReducer,
-  news: newsReducer,
-  selectedNews: selectedNewsReducer,
   races: racesReducer,
   selectedRace: selectedRaceReducer,
   alertsQueue: alertsQueueReducer
