+++ ui-web-react/src/model/sagas/sagas.ts
@@ -1,5 +1,4 @@
 import { all, call } from "redux-saga/effects";
-import { newsSaga, selectedNewsSaga } from "./news.saga";
 import { racesSaga, selectedRaceSaga } from "./race.saga";
 import {
   userLoginSaga,
@@ -14,9 +13,7 @@ export const delay = (ms: number) => new Promise(res => setTimeout(res, ms));
 
 function* raceLogSaga() {
   yield all([
-    call(newsSaga),
     call(racesSaga),
-    call(selectedNewsSaga),
     call(selectedRaceSaga),
     call(userLoginSaga),
     call(userLoginOnStartSaga),
